Never delete lexemes which are fragments of compound lexemes.

+++ wwwbase/admin/deTool.php
@@ -139,7 +139,8 @@ if ($saveButton) {
           $e->delete();
         }
       }
-      if (!$keepLexem) {
+      if (!$keepLexem &&
+          ($l->canDelete() == Lexem::CAN_DELETE_OK)) {
         $l->delete();
       }
     }
